new_typeform_pipeline/requirements.txt
openai>=1.0.0
requests>=2.25.0
google-api-python-client>=2.0.0
google-auth>=2.0.0
python-dotenv
flask 